Optimization of the Nearest Neighbor TSP Problem
This publication is dedicated to optimizing the nearest neighbor method to achieve better results for the overall TSP (traveling salesman problem) route. This optimization was designed not to significantly extend the time of this efficient method while achieving a significant improvement. The publication includes a reference to the algorithm of this improvement in the Python programming language. All aspects of this algorithm will be presented in the publication.